Move files into src
