+++ package.json
@@ -9,11 +9,12 @@
   },
   "dependencies": {
     "axios": "^0.21.1",
+    "ka-api": "github:smj-edison/ka-api",
     "node-diff3": "^2.1.0",
-    "proper-lockfile": "^4.1.1",
-    "qs": "^6.8.2",
+    "proper-lockfile": "^4.1.2",
+    "qs": "^6.9.6",
     "uuid": "^7.0.3",
-    "yargs": "^16.1.1"
+    "yargs": "^16.2.0"
   },
   "author": "Mason Jones",
   "license": "MIT"
