+++ src/command_error.js
