+++ src/command_router.js
@@ -7,7 +7,7 @@
 const test = require("./commands/test");
 const createprogram = require("./commands/createprogram");
 const deleteprogram = require("./commands/deleteprogram");
-const programsettings = require("./commands/programsettings.js");
+const programsettings = require("./src/commands/programsettings.js.js");
 const merge = require("./commands/merge/merge.js");
 const contrib = require("./commands/contrib");
