+++ src/commands/deleteprogram.js
