+++ src/commands/merge/merge.js
@@ -1,6 +1,6 @@
 const {updateProgramCodeAndHeaders, getProgramCodeAndHeaders, spinOffProgramCodeAndHeaders, deleteProgram, getProgramJSON} = require("../../programs");
 const {isAuthor, isContributor} = require("../../authorization");
-const {loadProgramHistory, updateProgramHistory} = require("../../program_history");
+const {loadProgramHistory, updateProgramHistory} = require("../../../program_history");
 const uuidv1 = require("uuid").v1;
 const Diff3 = require("node-diff3");
