+++ src/commands/test.js
