+++ src/discussion_culling.js
